CDH-45314: HIVE-4924: JDBC: Support query timeout for jdbc (Vaibhav Gumashta reviewed by Thejas Nair)
Conflicts:
	itests/hive-unit/src/test/java/org/apache/hive/jdbc/TestJdbcDriver2.java
	jdbc/src/java/org/apache/hive/jdbc/HiveStatement.java
	ql/src/java/org/apache/hadoop/hive/ql/Driver.java
	service-rpc/if/TCLIService.thrift
	service-rpc/src/gen/thrift/gen-cpp/TCLIService_types.cpp
	service-rpc/src/gen/thrift/gen-cpp/TCLIService_types.h
	service-rpc/src/gen/thrift/gen-javabean/org/apache/hive/service/rpc/thrift/TExecuteStatementReq.java
	service-rpc/src/gen/thrift/gen-javabean/org/apache/hive/service/rpc/thrift/TOperationState.java
	service-rpc/src/gen/thrift/gen-php/Types.php
	service-rpc/src/gen/thrift/gen-py/TCLIService/ttypes.py
	service-rpc/src/gen/thrift/gen-rb/t_c_l_i_service_types.rb
	service/src/java/org/apache/hive/service/cli/ICLIService.java
	service/src/java/org/apache/hive/service/cli/operation/MetadataOperation.java
	service/src/java/org/apache/hive/service/cli/operation/SQLOperation.java
	service/src/java/org/apache/hive/service/cli/session/HiveSessionImpl.java
	service/src/java/org/apache/hive/service/cli/thrift/RetryingThriftCLIServiceClient.java
	service/src/java/org/apache/hive/service/cli/thrift/ThriftCLIService.java
	service/src/test/org/apache/hive/service/cli/thrift/ThriftCliServiceTestWithCookie.java

+++ ql/src/java/org/apache/hadoop/hive/ql/session/OperationLog.java
@@ -152,7 +152,7 @@ public class OperationLog {
       return readResults(maxRows);
     }
 
-    void remove() {
+    synchronized void remove() {
       try {
         if (in != null) {
           in.close();
@@ -160,8 +160,10 @@ public class OperationLog {
         if (out != null) {
           out.close();
         }
-        FileUtils.forceDelete(file);
-        isRemoved = true;
+        if (!isRemoved) {
+          FileUtils.forceDelete(file);
+          isRemoved = true;
+        }
       } catch (Exception e) {
         LOG.error("Failed to remove corresponding log file of operation: " + operationName, e);
       }
